display-tracking-python/requirements.txt
altgraph==0.17.4
dnspython==2.7.0
googleapis-common-protos==1.66.0
grpclib==0.4.7
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
macholib==1.16.3
multidict==6.1.0
packaging==24.2
protobuf==5.29.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pymongo==4.11
setuptools==75.8.0
typing_extensions==4.12.2
viam-sdk==0.38.0
